+++ c/ndsort.c
@@ -35,6 +35,7 @@
 #include <getopt.h> // for getopt_long()
 #include <math.h>  // for INFINITY
 
+#include "common.h"
 #include "hv.h"
 #include "nondominated.h" // for normalise()
 
@@ -247,8 +248,8 @@ int main(int argc, char *argv[])
                        lower_range, upper_range,
                        lbound, ubound);
 
-            double *hvc = malloc (sizeof(double) * data_size);
-            hv_contributions (hvc, data, dim, data_size, ref);
+            double *hvc = malloc(sizeof(double) * data_size);
+            hv_contributions(hvc, data, dim, data_size, ref, /*ignore_dominated=*/true);
             /* FIXME: handle uevs: keep_uevs_flag ? uev : NULL);*/
             for (int k = 0, j = 0; k < size; k++) {
                 if (rank[k] != i) continue;